+++ src/waypoint_parser.cpp
@@ -25,7 +25,7 @@ void WaypointParser::parse(std::string json_path)
     using namespace boost::property_tree;
     ptree pt;
     read_json(json_path, pt);
-    boost::optional<uint8_t> start_waypoint_index = pt.get_optional<uint8_t>("start_waypoint_index");
+    boost::optional<int16_t> start_waypoint_index = pt.get_optional<int16_t>("start_waypoint_index");
     if(!start_waypoint_index)
     {
         ROS_ERROR_STREAM("failed to parse start_waypoint_index");
@@ -35,7 +35,7 @@ void WaypointParser::parse(std::string json_path)
     BOOST_FOREACH (const ptree::value_type& child, pt.get_child("waypoints"))
     {
         const ptree& waypoint = child.second;
-        boost::optional<uint8_t> index = waypoint.get_optional<uint8_t>("index");
+        boost::optional<int16_t> index = waypoint.get_optional<int16_t>("index");
         if(!index)
         {
             ROS_ERROR_STREAM("failed to parse waypoint index");
@@ -83,7 +83,7 @@ void WaypointParser::parse(std::string json_path)
             ROS_ERROR_STREAM("failed to parse frame_id");
             std::exit(0);
         }
-        boost::optional<std::vector<uint8_t> > next_waypoint_index = getNextWaypointIndex(waypoint);
+        boost::optional<std::vector<int16_t> > next_waypoint_index = getNextWaypointIndex(waypoint);
         if(!next_waypoint_index)
         {
             ROS_ERROR_STREAM("failed to parse next waypoint index");
@@ -99,14 +99,14 @@ void WaypointParser::parse(std::string json_path)
     }
 }
 
-boost::optional<std::vector<uint8_t> > WaypointParser::getNextWaypointIndex(boost::property_tree::ptree tree)
+boost::optional<std::vector<int16_t> > WaypointParser::getNextWaypointIndex(boost::property_tree::ptree tree)
 {
     using namespace boost::property_tree;
-    std::vector<uint8_t> ret;
+    std::vector<int16_t> ret;
     BOOST_FOREACH (const ptree::value_type& child, tree.get_child("next_waypoint_index"))
     {
         const ptree& p = child.second;
-        boost::optional<uint8_t> x = p.get_optional<uint8_t>("");
+        boost::optional<int16_t> x = p.get_optional<int16_t>("");
         if(!x)
         {
             return boost::none;
